+++ crates/file-serve/src/lib.rs
@@ -17,6 +17,8 @@
 //! ```
 
 #![cfg_attr(docsrs, feature(doc_auto_cfg))]
+#![warn(clippy::print_stderr)]
+#![warn(clippy::print_stdout)]
 
 use std::{
     str::FromStr,