+++ src/bin/cobalt/args.rs
@@ -1,20 +1,15 @@
-use std::env;
 use std::io::Write;
 use std::path;
 
 use anyhow::Context as _;
 
-use crate::error::*;
+use crate::error::Result;
 
 #[derive(Clone, Debug, PartialEq, Eq, clap::Args)]
-pub struct ConfigArgs {
+pub(crate) struct ConfigArgs {
     /// Config file to use [default: _cobalt.yml]
     #[arg(short, long, value_name = "FILE")]
     config: Option<path::PathBuf>,
-
-    /// Site destination folder [default: ./]
-    #[arg(short, long, value_name = "DIR")]
-    destination: Option<path::PathBuf>,
 
     /// Include drafts.
     #[arg(long)]
@@ -26,27 +21,17 @@ pub struct ConfigArgs {
 }
 
 impl ConfigArgs {
-    pub fn load_config(&self) -> Result<cobalt_config::Config> {
+    pub(crate) fn load_config(&self) -> Result<cobalt_config::Config> {
         let config_path = self.config.as_deref();
 
         // Fetch config information if available
         let mut config = if let Some(config_path) = config_path {
             cobalt_config::Config::from_file(config_path).with_context(|| {
-                anyhow::format_err!("Error reading config file {:?}", config_path)
+                anyhow::format_err!("Error reading config file {}", config_path.display())
             })?
         } else {
-            let cwd = env::current_dir().unwrap_or_default();
-            cobalt_config::Config::from_cwd(cwd)?
+            cobalt_config::Config::from_cwd(".")?
         };
-
-        config.abs_dest = self
-            .destination
-            .as_deref()
-            .map(|d| {
-                std::fs::create_dir_all(d)?;
-                dunce::canonicalize(d)
-            })
-            .transpose()?;
 
         if let Some(drafts) = self.drafts() {
             config.include_drafts = drafts;
@@ -55,7 +40,7 @@ impl ConfigArgs {
         Ok(config)
     }
 
-    pub fn drafts(&self) -> Option<bool> {
+    pub(crate) fn drafts(&self) -> Option<bool> {
         resolve_bool_arg(self.drafts, self.no_drafts)
     }
 }
@@ -69,17 +54,11 @@ fn resolve_bool_arg(yes: bool, no: bool) -> Option<bool> {
     }
 }
 
-pub fn init_logging(
+pub(crate) fn init_logging(
     level: clap_verbosity_flag::Verbosity<clap_verbosity_flag::InfoLevel>,
     colored: bool,
 ) {
     if let Some(level) = level.log_level() {
-        let palette = if colored {
-            Palette::colored()
-        } else {
-            Palette::plain()
-        };
-
         let mut builder = env_logger::Builder::new();
         builder.write_style(if colored {
             env_logger::WriteStyle::Always
@@ -92,32 +71,16 @@ pub fn init_logging(
         if level == log::LevelFilter::Trace {
             builder.format_timestamp_secs();
         } else {
-            builder.format(move |f, record| match record.level() {
-                log::Level::Error => writeln!(
-                    f,
-                    "{}: {}",
-                    palette.error.paint(record.level()),
-                    record.args()
-                ),
-                log::Level::Warn => writeln!(
-                    f,
-                    "{}: {}",
-                    palette.warn.paint(record.level()),
-                    record.args()
-                ),
-                log::Level::Info => writeln!(f, "{}", record.args()),
-                log::Level::Debug => writeln!(
-                    f,
-                    "{}: {}",
-                    palette.debug.paint(record.level()),
-                    record.args()
-                ),
-                log::Level::Trace => writeln!(
-                    f,
-                    "{}: {}",
-                    palette.trace.paint(record.level()),
-                    record.args()
-                ),
+            builder.format(move |f, record| {
+                let level = record.level();
+                let args = record.args();
+                match record.level() {
+                    log::Level::Error => writeln!(f, "{ERROR}{level}{ERROR:#}: {args}"),
+                    log::Level::Warn => writeln!(f, "{WARN}{level}{WARN:#}: {args}"),
+                    log::Level::Info => writeln!(f, "{args}"),
+                    log::Level::Debug => writeln!(f, "{DEBUG}{level}{DEBUG:#}: {args}"),
+                    log::Level::Trace => writeln!(f, "{TRACE}{level}{TRACE:#}: {args}"),
+                }
             });
         }
 
@@ -125,30 +88,7 @@ pub fn init_logging(
     }
 }
 
-#[derive(Copy, Clone, Debug)]
-struct Palette {
-    error: yansi::Style,
-    warn: yansi::Style,
-    debug: yansi::Style,
-    trace: yansi::Style,
-}
-
-impl Palette {
-    pub fn colored() -> Self {
-        Self {
-            error: yansi::Style::new(yansi::Color::Red).bold(),
-            warn: yansi::Style::new(yansi::Color::Yellow),
-            debug: yansi::Style::new(yansi::Color::Blue),
-            trace: yansi::Style::new(yansi::Color::Cyan),
-        }
-    }
-
-    pub fn plain() -> Self {
-        Self {
-            error: yansi::Style::default(),
-            warn: yansi::Style::default(),
-            debug: yansi::Style::default(),
-            trace: yansi::Style::default(),
-        }
-    }
-}
+const ERROR: anstyle::Style = anstyle::AnsiColor::Red.on_default().bold();
+const WARN: anstyle::Style = anstyle::AnsiColor::Yellow.on_default();
+const DEBUG: anstyle::Style = anstyle::AnsiColor::Blue.on_default();
+const TRACE: anstyle::Style = anstyle::AnsiColor::Cyan.on_default();