+++ src/bin/cobalt/build.rs
@@ -2,18 +2,31 @@ use std::env;
 use std::fs;
 
 use crate::args;
-use crate::error::*;
+use crate::error::Result;
 
 /// Build the cobalt project at the source dir
 #[derive(Clone, Debug, PartialEq, Eq, clap::Args)]
-pub struct BuildArgs {
+pub(crate) struct BuildArgs {
+    /// Site destination folder [default: ./_site]
+    #[arg(short, long, value_name = "DIR", help_heading = "Config")]
+    destination: Option<std::path::PathBuf>,
+
     #[command(flatten, next_help_heading = "Config")]
-    pub config: args::ConfigArgs,
+    pub(crate) config: args::ConfigArgs,
 }
 
 impl BuildArgs {
-    pub fn run(&self) -> Result<()> {
-        let config = self.config.load_config()?;
+    pub(crate) fn run(&self) -> Result<()> {
+        let mut config = self.config.load_config()?;
+        config.abs_dest = self
+            .destination
+            .as_deref()
+            .map(|d| {
+                fs::create_dir_all(d)?;
+                dunce::canonicalize(d)
+            })
+            .transpose()?;
+
         let config = cobalt::cobalt_model::Config::from_config(config)?;
 
         build(config)?;
@@ -23,7 +36,7 @@ impl BuildArgs {
     }
 }
 
-pub fn build(config: cobalt::Config) -> Result<()> {
+pub(crate) fn build(config: cobalt::Config) -> Result<()> {
     info!(
         "Building from `{}` into `{}`",
         config.source.display(),
@@ -36,13 +49,13 @@ pub fn build(config: cobalt::Config) -> Result<()> {
 
 /// Cleans `destination` directory
 #[derive(Clone, Debug, PartialEq, Eq, clap::Args)]
-pub struct CleanArgs {
+pub(crate) struct CleanArgs {
     #[command(flatten, next_help_heading = "Config")]
-    pub config: args::ConfigArgs,
+    pub(crate) config: args::ConfigArgs,
 }
 
 impl CleanArgs {
-    pub fn run(&self) -> Result<()> {
+    pub(crate) fn run(&self) -> Result<()> {
         let config = self.config.load_config()?;
         let config = cobalt::cobalt_model::Config::from_config(config)?;
 
@@ -50,7 +63,7 @@ impl CleanArgs {
     }
 }
 
-pub fn clean(config: &cobalt::Config) -> Result<()> {
+pub(crate) fn clean(config: &cobalt::Config) -> Result<()> {
     let cwd = env::current_dir().unwrap_or_default();
     let destdir = dunce::canonicalize(&config.destination);
     let destdir = match destdir {