+++ src/bin/cobalt/serve.rs
@@ -9,33 +9,33 @@ use notify::Watcher as _;
 
 use crate::args;
 use crate::build;
-use crate::error::*;
+use crate::error::Result;
 
 /// Build, serve, and watch the project at the source dir
 #[derive(Clone, Debug, PartialEq, Eq, clap::Args)]
-pub struct ServeArgs {
+pub(crate) struct ServeArgs {
     /// Open a browser
     #[arg(long)]
-    pub open: bool,
+    pub(crate) open: bool,
 
     /// Host to serve from
     #[arg(long, value_name = "HOSTNAME_OR_IP", default_value = "localhost")]
-    pub host: String,
+    pub(crate) host: String,
 
     /// Port to serve from
     #[arg(short = 'P', long, value_name = "NUM")]
-    pub port: Option<u16>,
+    pub(crate) port: Option<u16>,
 
     /// Disable rebuilding on change
     #[arg(long)]
-    pub no_watch: bool,
+    pub(crate) no_watch: bool,
 
     #[command(flatten, next_help_heading = "Config")]
-    pub config: args::ConfigArgs,
+    pub(crate) config: args::ConfigArgs,
 }
 
 impl ServeArgs {
-    pub fn run(&self) -> Result<()> {
+    pub(crate) fn run(&self) -> Result<()> {
         let dest = tempfile::tempdir()?;
 
         let mut server = file_serve::ServerBuilder::new(dest.path());
@@ -46,9 +46,9 @@ impl ServeArgs {
         let server = server.build();
 
         let mut config = self.config.load_config()?;
-        debug!("Overriding config `site.base_url` with `{}`", server.addr());
-        config.site.base_url = Some(format!("http://{}", server.addr()).into());
-        let mut config = cobalt::cobalt_model::Config::from_config(config)?;
+        debug!("Overriding config `site.base_url` with `/`");
+        config.site.base_url = Some("/".into());
+        let mut config = cobalt_model::Config::from_config(config)?;
         debug!(
             "Overriding config `destination` with `{}`",
             dest.path().display()
@@ -67,7 +67,7 @@ impl ServeArgs {
 
             dest.close()?;
         } else {
-            info!("Watching {:?} for changes", &config.source);
+            info!("Watching {} for changes", &config.source.display());
             thread::spawn(move || {
                 let e = serve(&server);
                 if let Some(e) = e.err() {
@@ -105,14 +105,19 @@ fn open_browser(url: String) -> Result<()> {
 fn watch(config: &cobalt_model::Config) -> Result<()> {
     // canonicalize is to ensure there is no question that `watcher`s paths come back safe for
     // Files::includes_file
-    let source = dunce::canonicalize(path::Path::new(&config.source))
-        .with_context(|| anyhow::format_err!("Failed in processing source"))?;
+    let source = dunce::canonicalize(path::Path::new(&config.source)).with_context(|| {
+        anyhow::format_err!("Failed in processing source `{}`", config.source.display())
+    })?;
 
     // Also canonicalize the destination folder. In particular for Windows, notify-rs
     // generates the absolute path by prepending the above source path.
     // On Windows canonicalize() adds a \\?\ to the start of the path.
-    let destination = dunce::canonicalize(&config.destination)
-        .with_context(|| anyhow::format_err!("Failed to canonicalize destination folder"))?;
+    let destination = dunce::canonicalize(&config.destination).with_context(|| {
+        anyhow::format_err!(
+            "Failed to canonicalize destination folder `{}`",
+            config.destination.display()
+        )
+    })?;
 
     let (tx, rx) = channel();
     let mut watcher =
@@ -120,7 +125,7 @@ fn watch(config: &cobalt_model::Config) -> Result<()> {
     watcher
         .watch(&source, notify::RecursiveMode::Recursive)
         .with_context(|| anyhow::format_err!("Notify error"))?;
-    info!("Watching {:?} for changes", &config.source);
+    info!("Watching {} for changes", config.source.display());
 
     for event in rx {
         let event = event.with_context(|| anyhow::format_err!("Notify error"))?;