+++ src/pagination/tags.rs
@@ -4,7 +4,7 @@ use crate::cobalt_model::pagination::PaginationConfig;
 use crate::cobalt_model::slug;
 use crate::document::Document;
 
-use super::*;
+use super::{create_all_paginators, helpers, paginator, sort_posts, Result, ValueView};
 use helpers::extract_tags;
 use paginator::Paginator;
 
@@ -20,7 +20,7 @@ fn distribute_posts_by_tags<'a>(
                     .ok_or_else(|| anyhow::format_err!("Should have string tags"))?
                     .to_kstr()
                     .into_string();
-                let cur_tag = per_tags.entry(tag).or_insert_with(Vec::new);
+                let cur_tag = per_tags.entry(tag).or_default();
                 cur_tag.push(post);
             }
         }
@@ -35,7 +35,7 @@ struct TagPaginators {
 }
 
 #[allow(clippy::bind_instead_of_map)]
-pub fn create_tags_paginators(
+pub(crate) fn create_tags_paginators(
     all_posts: &[&liquid::model::Value],
     doc: &Document,
     pagination_cfg: &PaginationConfig,