+++ CHANGELOG.md
@@ -7,6 +7,38 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 <!-- next-header -->
 ## [Unreleased] - ReleaseDate
 
+## [0.19.5] - 2024-07-05
+
+### Fixes
+
+- *(build)* List correct default for `--destination`
+- *(serve)* Remove unused `--destination`
+
+## [0.19.4] - 2024-07-05
+
+## [0.19.3] - 2024-04-15
+
+### Compatibility
+
+- MSRV is now 1.76.0
+
+### Fixes
+
+- `cobalt serve -c <file>` should now work
+- Cleaned up error messages
+
+## [0.19.2] - 2024-02-28
+
+## [0.19.1] - 2024-02-28
+
+### Compatibility
+
+- MSRV is now 1.73
+
+### Fixes
+
+- Put the base url for `cobalt serve` at `/` to work with tunnelling
+
 ## [0.19.0] - 2023-06-06
 
 ### Compatibility
@@ -683,7 +715,12 @@ this.
 [@tak1n]: https://github.com/tak1n
 
 <!-- next-url -->
-[Unreleased]: https://github.com/cobalt-org/cobalt.rs/compare/v0.19.0...HEAD
+[Unreleased]: https://github.com/cobalt-org/cobalt.rs/compare/v0.19.5...HEAD
+[0.19.5]: https://github.com/cobalt-org/cobalt.rs/compare/v0.19.4...v0.19.5
+[0.19.4]: https://github.com/cobalt-org/cobalt.rs/compare/v0.19.3...v0.19.4
+[0.19.3]: https://github.com/cobalt-org/cobalt.rs/compare/v0.19.2...v0.19.3
+[0.19.2]: https://github.com/cobalt-org/cobalt.rs/compare/v0.19.1...v0.19.2
+[0.19.1]: https://github.com/cobalt-org/cobalt.rs/compare/v0.19.0...v0.19.1
 [0.19.0]: https://github.com/cobalt-org/cobalt.rs/compare/v0.18.5...v0.19.0
 [0.18.5]: https://github.com/cobalt-org/cobalt.rs/compare/v0.18.4...v0.18.5
 [0.18.4]: https://github.com/cobalt-org/cobalt.rs/compare/v0.18.3...v0.18.4