+++ crates/file-serve/CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 <!-- next-header -->
 ## [Unreleased] - ReleaseDate
 
+## [0.3.4] - 2024-07-05
+
+## [0.3.3] - 2024-07-05
+
+### Compatibility
+
+- MSRV is now 1.76
+
+## [0.3.2] - 2024-04-15
+
+## [0.3.1] - 2024-02-28
+
+### Compatibility
+
+- MSRV is now 1.73
+
 ## [0.3.0] - 2023-08-22
 
 ### Breaking Change
@@ -42,7 +58,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.1.1] - 2022-04-04
 
 <!-- next-url -->
-[Unreleased]: https://github.com/cobalt-org/cobalt.rs/compare/file-serve-v0.3.0...HEAD
+[Unreleased]: https://github.com/cobalt-org/cobalt.rs/compare/file-serve-v0.3.4...HEAD
+[0.3.4]: https://github.com/cobalt-org/cobalt.rs/compare/file-serve-v0.3.3...file-serve-v0.3.4
+[0.3.3]: https://github.com/cobalt-org/cobalt.rs/compare/file-serve-v0.3.2...file-serve-v0.3.3
+[0.3.2]: https://github.com/cobalt-org/cobalt.rs/compare/file-serve-v0.3.1...file-serve-v0.3.2
+[0.3.1]: https://github.com/cobalt-org/cobalt.rs/compare/file-serve-v0.3.0...file-serve-v0.3.1
 [0.3.0]: https://github.com/cobalt-org/cobalt.rs/compare/file-serve-v0.2.4...file-serve-v0.3.0
 [0.2.4]: https://github.com/cobalt-org/cobalt.rs/compare/file-serve-v0.2.3...file-serve-v0.2.4
 [0.2.3]: https://github.com/cobalt-org/cobalt.rs/compare/file-serve-v0.2.2...file-serve-v0.2.3